move 'locationCodes' and 'appConfig' to 'src/app/data'; modify rel paths
