+++ src/app/components/HoldRequest/HoldRequest.jsx
@@ -13,7 +13,7 @@ import DocumentTitle from 'react-document-title';
 import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
 
 import PatronStore from '../../stores/PatronStore';
-import appConfig from '../../appConfig';
+import appConfig from '../../data/appConfig';
 import LibraryItem from '../../utils/item';
 import LoadingLayer from '../LoadingLayer/LoadingLayer';
 import { trackDiscovery } from '../../utils/utils';
