+++ src/app/components/Search/Search.jsx
@@ -7,7 +7,7 @@ import {
   trackDiscovery,
   ajaxCall,
 } from '../../utils/utils';
-import appConfig from '../../appConfig';
+import appConfig from '../../data/appConfig';
 
 /**
  * The main container for the top Search section.
