+++ src/server/ApiRoutes/Hold.js
@@ -4,8 +4,8 @@ import {
   omit as _omit,
 } from 'underscore';
 
-import appConfig from '../../app/appConfig';
-import locationCodes from '../../../locationCodes';
+import appConfig from '../../app/data/appConfig';
+import locationCodes from '../app/data/locationCodes';
 import locationDetails from '../../../locations';
 import User from './User';
 import Bib from './Bib';
